+++ tests/testthat/test-queue.R
@@ -20,11 +20,10 @@ test_that("queue works as intended", {
   expect_equal(queue$queue$worker_log_tail(worker_2, 3)[3, "message"],
                "TIMEOUT_SET")
 
-
   expect_length(queue$queue$task_list(), 0)
 
-  ## jobs can be pushed to queue
-  job_id <- queue$submit(NULL, list())
+  ## model run can be pushed to queue
+  job_id <- queue$submit_model_run(NULL, list())
   expect_length(queue$queue$task_list(), 1)
 
   ## status can be retireved
@@ -66,6 +65,41 @@ test_that("queue works as intended", {
   expect_equal(con$SCARD(key), 0)
 })
 
+test_that("queue can run arbitrary jobs", {
+  test_redis_available()
+
+  queue <- Queue$new(timeout = 300)
+
+  ## model run can be pushed to queue
+  job_id <- queue$submit(quote({
+    Sys.sleep(1)
+    1 + 1
+  }))
+  expect_length(queue$queue$task_list(), 1)
+
+  ## status can be retireved
+  Sys.sleep(0.1)
+  status <- queue$status(job_id)
+  expect_equal(status$status, "RUNNING")
+  expect_false(status$done)
+  expect_equal(status$success, json_verbatim("null"))
+  expect_equal(status$queue, 0)
+  expect_null(status$progress)
+
+  ## After task has completed
+  result <- queue$queue$task_wait(job_id)
+  status <- queue$status(job_id)
+  expect_equal(status$status, "COMPLETE")
+  expect_true(status$done)
+  expect_true(status$success)
+  expect_equal(status$queue, 0)
+  expect_null(status$progress)
+
+  ## Result can be retrieved after task has completed
+  res <- queue$result(job_id)
+  expect_equal(res, 2)
+})
+
 test_that("queue_id is generated if not supplied", {
   withr::with_envvar(
     c("HINTR_QUEUE_ID" = NA),
@@ -98,7 +132,6 @@ test_that("test queue starts workers with timeout", {
   expect_equal(timeout[[2]][["timeout"]], 300.0)
 })
 
-
 test_that("queue starts up normally without a timeout", {
   queue <- Queue$new(workers = 1)
   on.exit(queue$cleanup())
@@ -106,4 +139,39 @@ test_that("queue starts up normally without a timeout", {
                                                queue$queue$worker_list(),
                                                progress = FALSE)
   expect_equal(timeout[[1]], c("timeout" = Inf, remaining = Inf))
+})
+
+test_that("queue object starts up 2 queues", {
+  queue <- test_queue(workers = 2)
+  expect_equal(queue$queue$worker_config_read("localhost")$queue,
+               c(QUEUE_CALIBRATE, QUEUE_RUN))
+  queue$submit(quote(sin(1)), queue = QUEUE_CALIBRATE)
+  run_id <- queue$submit(quote(sin(1)), queue = QUEUE_RUN)
+  other_id <- queue$submit(quote(sin(1)), queue = "other")
+  queue$queue$task_wait(run_id)
+  expect_equal(queue$queue$queue_list(QUEUE_RUN), character(0))
+  expect_equal(queue$queue$queue_list(QUEUE_CALIBRATE), character(0))
+  ## Task submitted to "other" never gets run because this queue isn't run
+  ## by workers.
+  expect_equal(queue$queue$queue_list("other"), other_id)
+})
+
+test_that("calibrate gets run before model running", {
+  queue <- test_queue(workers = 0)
+  worker <- create_blocking_worker(queue$queue$queue_id)
+  run_id <- queue$submit_model_run(NULL, NULL)
+  ## Calibrate tasks will error but that is fine - we want to test here
+  ## that calibrate & model run get queued and run in the correct order
+  calibrate_id <- queue$submit_calibrate(NULL, NULL)
+
+  expect_equal(unname(queue$queue$task_status(c(run_id, calibrate_id))),
+               rep("PENDING", 2))
+  expect_equal(queue$queue$queue_list(QUEUE_RUN), run_id)
+  expect_equal(queue$queue$queue_list(QUEUE_CALIBRATE), calibrate_id)
+  worker$step(TRUE)
+  expect_equal(unname(queue$queue$task_status(c(run_id, calibrate_id))),
+               c("PENDING", "ERROR"))
+  worker$step(TRUE)
+  expect_equal(unname(queue$queue$task_status(c(run_id, calibrate_id))),
+               c("COMPLETE", "ERROR"))
 })